feat(features): Context 기능 로직 features 레이어로 분리
- features/auth/model: AppProvider, useApp
- features/chat/model: ChatProvider, useChat
- features/calendar/model: EventsProvider, useEvents
- features/tasks/model: TasksProvider, useTasks
- features/attendance/ui: AnimatedClockRing, ClockTimePicker, SetWorkDaysPersonal
- context/ 파일은 re-export 스텁으로 교체 (하위 호환 유지)
- App.tsx 및 pages/*.tsx import 경로 features로 업데이트

closes #4
