+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom'
-import { AppProvider } from './context/AppContext'
-import { TasksProvider } from './context/TasksContext'
-import { EventsProvider } from './context/EventsContext'
-import { ChatProvider } from './context/ChatContext'
+import { AppProvider } from './features/auth/model'
+import { TasksProvider } from './features/tasks/model'
+import { EventsProvider } from './features/calendar/model'
+import { ChatProvider } from './features/chat/model'
 import { Layout } from './components/Layout'
 import { Dashboard } from './pages/Dashboard'
 import { Chat } from './pages/Chat'
