+++ src/features/attendance/ui/ClockTimePicker.css
@@ -1,0 +1,102 @@
+.clock-time-picker {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 12px;
+}
+
+.clock-face {
+  width: 160px;
+  height: 160px;
+  border-radius: 50%;
+  background: rgba(15, 13, 26, 0.6);
+  border: 2px solid rgba(148, 163, 184, 0.2);
+  position: relative;
+  flex-shrink: 0;
+}
+
+.clock-mark {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  width: 24px;
+  height: 24px;
+  margin-left: -12px;
+  margin-top: -12px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 0.75rem;
+  color: var(--text-secondary);
+  transform-origin: 12px 12px;
+}
+
+.clock-hand {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform-origin: left center;
+  cursor: grab;
+  transition: transform 0.05s ease-out;
+}
+
+.clock-hand:active,
+.clock-hand.active {
+  cursor: grabbing;
+}
+
+.hour-hand {
+  width: 35px;
+  height: 4px;
+  margin-top: -2px;
+  background: var(--text-primary);
+  border-radius: 2px;
+  box-shadow: 0 0 4px rgba(0, 0, 0, 0.3);
+}
+
+.minute-hand {
+  width: 50px;
+  height: 3px;
+  margin-top: -1.5px;
+  background: var(--accent-purple);
+  border-radius: 2px;
+  box-shadow: 0 0 4px rgba(124, 58, 237, 0.4);
+}
+
+.clock-center {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  width: 10px;
+  height: 10px;
+  margin-left: -5px;
+  margin-top: -5px;
+  border-radius: 50%;
+  background: var(--accent-purple);
+}
+
+.clock-display {
+  font-size: 0.95rem;
+  font-weight: 600;
+  color: var(--text-primary);
+}
+
+.clock-ampm-toggle {
+  display: flex;
+  gap: 8px;
+}
+
+.clock-ampm-toggle button {
+  padding: 6px 14px;
+  border-radius: 8px;
+  font-size: 0.8rem;
+  background: rgba(148, 163, 184, 0.1);
+  color: var(--text-secondary);
+  border: 1px solid rgba(148, 163, 184, 0.2);
+}
+
+.clock-ampm-toggle button.active {
+  background: var(--accent-purple);
+  color: var(--text-primary);
+  border-color: var(--accent-purple);
+}
