+++ src/features/calendar/model/index.ts
@@ -1,0 +1,2 @@
+export { EventsProvider, useEvents } from './EventsProvider'
+export type { CalendarEvent } from './EventsProvider'
