+++ src/features/chat/model/index.ts
@@ -1,0 +1,2 @@
+export { ChatProvider, useChat } from './ChatProvider'
+export type { ChatMessage, Channel } from './ChatProvider'
