+++ src/features/tasks/model/index.ts
@@ -1,0 +1,3 @@
+export { TasksProvider, useTasks } from './TasksProvider'
+export type { Task, TaskPriority } from './TasksProvider'
+export { getTodayStr, formatDateDisplay } from './TasksProvider'
