+++ src/pages/Calendar.tsx
@@ -9,10 +9,10 @@ import koLocale from '@fullcalendar/core/locales/ko'
 import type { EventClickArg, EventDropArg, EventInput } from '@fullcalendar/core'
 import type { DateClickArg, EventResizeDoneArg } from '@fullcalendar/interaction'
 import { TimeInput } from '../components/TimeInput'
-import { useApp } from '../context/AppContext'
-import { useTasks, type Task, type TaskPriority } from '../context/TasksContext'
-import { useEvents, type CalendarEvent } from '../context/EventsContext'
-import { formatDateDisplay, getTodayStr } from '../context/TasksContext'
+import { useApp } from '../features/auth/model'
+import { useTasks, type Task, type TaskPriority } from '../features/tasks/model'
+import { useEvents, type CalendarEvent } from '../features/calendar/model'
+import { formatDateDisplay, getTodayStr } from '../features/tasks/model'
 import './Calendar.css'
 
 const PRIORITY_LABELS: Record<TaskPriority, string> = {
